Part of user profile related data CRUD added

+++ server/src/services/commentService.ts
@@ -26,8 +26,9 @@ export const editComment = async (text: string, commentId: string) => {
 }
 
 export const removeComment = async (commentId: string, postId: string) => {
-    await Comment.deleteOne({_id: commentId});
-    console.log(commentId);
-    await Like.deleteMany({_targetId: commentId});
-    await adjustCommentCount(postId, -1);
+    const result = await Comment.deleteOne({_id: commentId});
+    if (result.deletedCount === 1) {
+        await Like.deleteMany({_targetId: commentId});
+        await adjustCommentCount(postId, -1);
+    }
 }
